+++ src/routes/+page.svelte
@@ -1,2 +1,9 @@
-<h1>Welcome to SvelteKit</h1>
-<p>Visit <a href="https://svelte.dev/docs/kit">svelte.dev/docs/kit</a> to read the documentation</p>
+<script lang="ts">
+	import PlayerShell from '$lib/components/player-shell.svelte';
+</script>
+
+<svelte:head>
+	<title>CDP-XA7ES</title>
+</svelte:head>
+
+<PlayerShell />